MyDailyReminder-Backend/requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
DateTime==5.5 
gunicorn==23.0.0
pymongo==4.10.1